requirements.txt
# Core dependencies
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy==1.24.3

# Audio processing
pyttsx3==2.90
pygame==2.5.2

# AI/ML
openai==1.3.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
